Basic functionality, anti-aliasing needs fixing up
Tool reads the plat lines from a G* file and creates them on a grey copy
of the street - pixel by pixel. All trees are then placed using spice
trees so that it is clear whther plat lines are too close to the base of
the tree.

The output report indicates if the line crosses the tree image. Plat
line changes from red to purple to show the overlap also.

As the line is a bit clunky, need to try an anti-aliasing to see if it
gives a better result. May not be worth the trouble.