Use street snap as background rather than grey
Rather than using a grey street, a street snap is used as the background
- so will be able to see why a plat line is where it is. The overlaid
spice tree is now purple, so it stands out as an obvious addition to the
image.

If a valid street snap is not found, then the default grey street will
be used, with the original spice tree image.