Fixed bug caused by overlapping tree canopies
For trees on a grey street, the platline was determined to cross a tree
root if the constructed street image tree didn't match the reference
tree sample. However when two trees were close together, the canopy of
the second tree would overlay the canopy of the first - and so this
would cause a tree to be marked as a fail, when it was simply the
canopies overlapping. As a different test is made for case when street
snaps, this explained why I was seeing different numbers of good/bad
trees depending on whether I produced grey/street snap streets.